fix: ASIO Segfault on Windows
This commit fixes a bug that occurs in the ASIO host on windows. It
commonly manifests in a message about Data Execution Prevention.
Essentially, the ASIO API call ASIOCreateBuffers() expects the callbacks
struct to exist for the lifetime of the buffers (i.e. until
ASIOFreeBuffers() is called). This isn't really mentioned in the ASIO
Docs unfortunately, and I only found it by hunting for this bug.

This commit fixes the issue by making the AsioCallbacks struct a static
object rather than a local that is constructed by a function. It has to
be mutable, unfortunately, because the C api for ASIOCreateBuffers()
asks for a non-const pointer. I can't think of any reason that it would
be mutated by the driver, so I still think this is the best solution.